修改部分代码 添加护理病历模块（旧版XML解析）

+++ libraries/dhcc_infusion/src/main/java/com/dhcc/module/infusion/view/CustomBarChart.java
@@ -133,7 +133,7 @@ public class CustomBarChart extends LinearLayout {
      */
     public void setMax(int max) {
         this.MAX = max;
-        String i = max > 1 ? "" + (int) max / 2 : "0.5";
+        String i = max > 1 ? "" + max / 2 : "0.5";
         tvSpliceLine.setText(i);
     }
